requirements.txt
pywebview>=4.0
pillow>=10.0.0
mss>=9.0.0
pynput>=1.7.6
pywin32>=305